simplify code / respond to a few nits
- more accurate name for return value of `undelegate` function (`queuedWithdrawal` => `withdrawalRoot`)
- absorb internal `_undelegate` function into the external `delegate` function, since it was now being called in only one place
- mark a few uints explicitly as uint256s

+++ src/test/utils/Operators.sol
@@ -60,11 +60,11 @@ contract Operators is Test {
         return pubkey;
     }
 
-    function readUint(string memory json, uint256 index, string memory key) public returns (uint) {
+    function readUint(string memory json, uint256 index, string memory key) public returns (uint256) {
         return stringToUint(stdJson.readString(json, string.concat(operatorPrefix(index), key)));
     }
 
-    function stringToUint(string memory s) public pure returns (uint) {
+    function stringToUint(string memory s) public pure returns (uint256) {
         bytes memory b = bytes(s);
         uint256 result = 0;
         for (uint256 i = 0; i < b.length; i++) {
